+++ src/tooling/piral-cli/src/common/importmap.test.ts
@@ -2,7 +2,13 @@ import { readImportmap } from './importmap';
 
 jest.mock('./npm', () => ({
   tryResolvePackage(id, dir) {
-    return `${dir}/${id}/index.js`;
+    if (id === 'qxz') {
+      return undefined;
+    } else if (id.endsWith('package.json')) {
+      return `${dir}/${id}`;
+    } else {
+      return `${dir}/${id}/index.js`;
+    }
   },
 }));
 
@@ -14,6 +20,28 @@ const mockPackages = {
   '/data/bar': {
     name: 'bar',
     version: '1.0.0',
+  },
+  '/data/app1': {
+    name: 'app1',
+    version: '1.0.0',
+    importmap: {
+      imports: {
+        'foo@1.2.3': 'foo',
+      },
+    },
+  },
+  '/data/app1/foo': {
+    name: 'foo',
+    version: '1.2.3',
+  },
+  '/data/app2': {
+    name: 'app1',
+    version: '1.0.0',
+    importmap: {
+      imports: {
+        'qxz@1.0.0': '',
+      },
+    },
   },
 };
 
@@ -33,28 +61,20 @@ jest.mock('./io', () => ({
 
 describe('Importmap', () => {
   it('reads empty one from package.json', async () => {
-    const deps = await readImportmap(
-      '/data',
-      {
-        importmap: {},
-      },
-      false,
-    );
+    const deps = await readImportmap('/data', {
+      importmap: {},
+    });
     expect(deps).toEqual([]);
   });
 
   it('reads fully qualified local dependency', async () => {
-    const deps = await readImportmap(
-      '/data',
-      {
-        importmap: {
-          imports: {
-            'foo@1.2.3': 'foo',
-          },
+    const deps = await readImportmap('/data', {
+      importmap: {
+        imports: {
+          'foo@1.2.3': 'foo',
         },
       },
-      false,
-    );
+    });
     expect(deps).toEqual([
       {
         alias: undefined,
@@ -70,17 +90,13 @@ describe('Importmap', () => {
   });
 
   it('reads fully qualified local dependency with implied exact version', async () => {
-    const deps = await readImportmap(
-      '/data',
-      {
-        importmap: {
-          imports: {
-            foo: 'foo',
-          },
+    const deps = await readImportmap('/data', {
+      importmap: {
+        imports: {
+          foo: 'foo',
         },
       },
-      false,
-    );
+    });
     expect(deps).toEqual([
       {
         alias: undefined,
@@ -105,7 +121,6 @@ describe('Importmap', () => {
           },
         },
       },
-      false,
       'match-major',
     );
     expect(deps).toEqual([
@@ -132,7 +147,6 @@ describe('Importmap', () => {
           },
         },
       },
-      false,
       'any-patch',
     );
     expect(deps).toEqual([
@@ -159,7 +173,6 @@ describe('Importmap', () => {
           },
         },
       },
-      false,
       'all',
     );
     expect(deps).toEqual([
@@ -178,17 +191,57 @@ describe('Importmap', () => {
 
   it('fails when the local version does not match the desired version', async () => {
     await expect(
-      readImportmap(
-        '/data',
-        {
-          importmap: {
-            imports: {
-              'bar@^2.0.0': '',
-            },
+      readImportmap('/data', {
+        importmap: {
+          imports: {
+            'bar@^2.0.0': '',
           },
         },
-        false,
-      ),
+      }),
     ).rejects.toThrow();
+  });
+
+  it('fails when the explicitly shared package is not installed', async () => {
+    await expect(
+      readImportmap('/data', {
+        importmap: {
+          imports: {
+            qxz: '',
+          },
+        },
+      }),
+    ).rejects.toThrow();
+  });
+
+  it('accepts an importmap with valid resolvable inherited deps', async () => {
+    const deps = await readImportmap('/data', {
+      importmap: {
+        imports: {},
+        inherit: ['app1'],
+      },
+    });
+    expect(deps).toEqual([
+      {
+        alias: undefined,
+        entry: '/data/app1/foo/index.js',
+        id: 'foo@1.2.3',
+        isAsync: false,
+        name: 'foo',
+        ref: 'foo.js',
+        requireId: 'foo@1.2.3',
+        type: 'local',
+        parents: ['app1'],
+      },
+    ]);
+  });
+
+  it('accepts an importmap with valid but unresolvable inherited deps', async () => {
+    const deps = await readImportmap('/data', {
+      importmap: {
+        imports: {},
+        inherit: ['app2'],
+      },
+    });
+    expect(deps).toEqual([]);
   });
 });