+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Piral Changelog
 
+## 1.3.1 (tbd)
+
+- Updated behavior with unresolved inherited dependencies (#633)
+- Updated behavior to only update installed dependencies on `pilet upgrade`
+- Added check to prevent installing of invalid versions locally (#635)
+
 ## 1.3.0 (October 9, 2023)
 
 - Fixed issue with global installation in pnpm (#624)